+++ src/Logitar.Portal.Core/Caching/ICacheService.cs
@@ -1,0 +1,8 @@
+﻿using Logitar.Portal.Core.Realms;
+
+namespace Logitar.Portal.Core.Caching;
+
+public interface ICacheService
+{
+  RealmAggregate? PortalRealm { get; set; }
+}