+++ src/Logitar.Portal.Core/Configurations/Commands/InitializeConfigurationHandler.cs
@@ -1,4 +1,5 @@
-﻿using Logitar.Portal.Core.Realms;
+﻿using Logitar.Portal.Core.Caching;
+using Logitar.Portal.Core.Realms;
 using Logitar.Portal.Core.Users;
 using Logitar.Portal.Core.Users.Contact;
 using MediatR;
@@ -6,26 +7,28 @@ using System.Globalization;
 
 namespace Logitar.Portal.Core.Configurations.Commands;
 
-internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfiguration>
+internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfiguration, Unit>
 {
   private const string PortalDisplayName = "Portal";
   private const string PortalDescription = "The realm in which the administrator users of the Portal belong to.";
 
+  private readonly ICacheService _cacheService;
   private readonly ICurrentActor _currentActor;
   private readonly IRealmRepository _realmRepository;
   private readonly IUserRepository _userRepository;
 
-  public InitializeConfigurationHandler(
+  public InitializeConfigurationHandler(ICacheService cacheService,
     ICurrentActor currentActor,
     IRealmRepository realmRepository,
     IUserRepository userRepository)
   {
+    _cacheService = cacheService;
     _currentActor = currentActor;
     _realmRepository = realmRepository;
     _userRepository = userRepository;
   }
 
-  public async Task Handle(InitializeConfiguration request, CancellationToken cancellationToken)
+  public async Task<Unit> Handle(InitializeConfiguration request, CancellationToken cancellationToken)
   {
     InitializeConfigurationInput input = request.Input;
     CultureInfo defaultLocale = input.DefaultLocale.GetRequiredCultureInfo(nameof(input.DefaultLocale));
@@ -47,6 +50,8 @@ internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfig
         customAttributes: customAttributes);
 
       await _realmRepository.SaveAsync(realm, cancellationToken);
+
+      _cacheService.PortalRealm = realm;
     }
 
     InitialUserInput userInput = input.User;
@@ -56,5 +61,7 @@ internal class InitializeConfigurationHandler : IRequestHandler<InitializeConfig
     user.SetEmail(_currentActor.Id, new ReadOnlyEmail(userInput.EmailAddress));
 
     await _userRepository.SaveAsync(user, cancellationToken);
+
+    return Unit.Value;
   }
 }