+++ src/Logitar.Portal.Core/IApplicationContext.cs
@@ -2,5 +2,7 @@
 
 public interface IApplicationContext
 {
+  Guid? ActivityId { get; set; }
+
   Uri? BaseUrl { get; }
 }