+++ src/Logitar.Portal.EntityFrameworkCore.PostgreSQL/Migrations/20230413191033_CreateActivityTable.cs
@@ -1,0 +1,67 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
+
+#nullable disable
+
+namespace Logitar.Portal.EntityFrameworkCore.PostgreSQL.Migrations
+{
+    /// <inheritdoc />
+    public partial class CreateActivityTable : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "Activities",
+                schema: "Logging",
+                columns: table => new
+                {
+                    ActivityId = table.Column<long>(type: "bigint", nullable: false)
+                        .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
+                    Id = table.Column<Guid>(type: "uuid", nullable: false),
+                    LogId = table.Column<long>(type: "bigint", nullable: false),
+                    Type = table.Column<string>(type: "character varying(255)", maxLength: 255, nullable: false),
+                    Data = table.Column<string>(type: "jsonb", nullable: false),
+                    StartedOn = table.Column<DateTime>(type: "timestamp with time zone", nullable: false),
+                    EndedOn = table.Column<DateTime>(type: "timestamp with time zone", nullable: true),
+                    Duration = table.Column<TimeSpan>(type: "interval", nullable: true),
+                    IsCompleted = table.Column<bool>(type: "boolean", nullable: false),
+                    Level = table.Column<string>(type: "character varying(255)", maxLength: 255, nullable: false),
+                    Errors = table.Column<string>(type: "jsonb", nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Activities", x => x.ActivityId);
+                    table.ForeignKey(
+                        name: "FK_Activities_Logs_LogId",
+                        column: x => x.LogId,
+                        principalSchema: "Logging",
+                        principalTable: "Logs",
+                        principalColumn: "LogId",
+                        onDelete: ReferentialAction.Cascade);
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Activities_Id",
+                schema: "Logging",
+                table: "Activities",
+                column: "Id",
+                unique: true);
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Activities_LogId",
+                schema: "Logging",
+                table: "Activities",
+                column: "LogId");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "Activities",
+                schema: "Logging");
+        }
+    }
+}