+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enhanced realm JWT secret.
 - Reimplemented password recovery.
 - Implemented Basic authentication.
+- Reimplemented logging.
 
 ## [1.1.5] - 2022-10-27